created nav wrapper component
