+++ src/router/index.ts
@@ -6,22 +6,19 @@ import TabsPage from '../views/TabsPage.vue';
 const routes: Array<RouteRecordRaw> = [
   {
     path: '/',
-    component: TabsPage,
     redirect: '/feed',
-    children: [
-      {
-        path: '/feed',
-        component: () => import('@/views/FeedTab.vue')
-      },
-      {
-        path: '/create_post',
-        component: () => import('@/views/CreatePostTab.vue')
-      },
-      {
-        path: '/profile',
-        component: () => import('@/views/ProfileTab.vue')
-      },
-    ]
+  },
+  {
+    path: '/feed',
+    component: () => import('@/views/FeedTab.vue')
+  },
+  {
+    path: '/create_post',
+    component: () => import('@/views/CreatePostTab.vue')
+  },
+  {
+    path: '/profile',
+    component: () => import('@/views/ProfileTab.vue')
   },
   {
     path: '/signIn',
@@ -30,10 +27,6 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/createAccount',
     component: () => import('@/views/CreateAccount.vue')
-  },
-  {
-    path: '/test',
-    component: () => import('@/components/Navigation/NavigationBottomBar.vue')
   }
 ]
